// src/app/modules/main/dtos/series-dto.d.ts
export interface SeriesDto {
    id: number;
    iddeclaration : number, 
    idarancel : number, 
    idpais : number,
    codigo_liberacion : string,
    fob : number, 
    fob_expenses : number,
    flete : number,
    sure : number,
    cif : number,
    kilos : number,
    quantity : number,
    units : number,
    idunittype : number,
    business_unit : number, 
    restring : number, 
    quantity_risk : number,
    preceding : string, 
    authorization : string, 
    idstatuscommodity : number,
    transport_document : string,
    name : string,
    mark : string,
    model : string,
    description : string,
    url : string,
}